Generalization FlowOverview first Implementation
